+++ src/webviews/rebase/protocol.ts
@@ -43,11 +43,9 @@ export interface Commit {
 // COMMANDS
 
 export const AbortCommandType = new IpcCommandType('rebase/abort');
-
 export const DisableCommandType = new IpcCommandType('rebase/disable');
-
+export const SearchCommandType = new IpcCommandType('rebase/search');
 export const StartCommandType = new IpcCommandType('rebase/start');
-
 export const SwitchCommandType = new IpcCommandType('rebase/switch');
 
 export interface ReorderParams {